Added sqlite to the bunch.

+++ auxiliary/output_streamer.hpp
@@ -198,7 +198,7 @@ private:
          stream << " --";
       // End of the line if not requested otherwise
       if (testTrait(rewrite_ln, trait_mask) && !testTrait(no_out, trait_mask))
-         stream << string(max(0u, PAD_LENGTH - string(stream_data).size()), ' ');
+         stream << string(max(static_cast<size_t>(0), PAD_LENGTH - string(stream_data).size()), ' ');
       if (!testTrait(no_newl, trait_mask))
          stream << endl;
    }
